Extract code related to the current Dedukti theory from CommonConcepts.hs
