+++ src/CommonConcepts.hs
@@ -3,14 +3,11 @@
 
 module CommonConcepts where
 
-import Dedukti.AbsDedukti
 import Utils
 import qualified Data.Map as M
 import qualified Data.Set as S
 import Data.List (isSuffixOf)
 import Data.Char
-
-type DTree a = Dedukti.AbsDedukti.Tree a
 
 -- Informath category symbols with arities
 
@@ -91,75 +88,6 @@ proofCats :: S.Set SCat
 proofCats = S.fromList ["Label", "Proof", "ProofExp", "Unit"]
 
 
--- referring to BaseConstants.dk
-
-identFalse = QIdent "false"
-identConj = QIdent "and"
-identDisj = QIdent "or"
-identImpl = QIdent "if"
-identPi = QIdent "forall"
-identSigma = QIdent "exists"
-identNot = QIdent "not"
-identEquiv = QIdent "iff"
-
-identDig = QIdent "Dig"
-identNat =  QIdent "Nat"
-identInt =  QIdent "Int"
-identRat =  QIdent "Rat"
-identReal =  QIdent "Real"
-identnat2int = QIdent "nat2int"
-identNat2real = QIdent "nat2real"
-identInt2real = QIdent "int2real"
-identRat2real = QIdent "rat2real"
-identPlus =  QIdent "plus"
-identMinus =  QIdent "minus"
-identTimes =  QIdent "times"
-identDiv =  QIdent "div"
-identPow = QIdent "pow"
-identNeg = QIdent "neg"
-identEq =  QIdent "Eq"
-identLt =  QIdent "Lt"
-identGt =  QIdent "Gt"
-identNeq =  QIdent "Neq"
-identLeq =  QIdent "Leq"
-identGeq =  QIdent "Geq"
-
--- Peano-style Nat constructors in BaseConstants.dk
-identZero = QIdent "0"
-identSucc = QIdent "succ"
-
--- these are to be peeled away
-dkProof = "Proof"
-dkElem  = "Elem"
-identProof = QIdent dkProof
-identElem = QIdent dkElem
-
-identSuchThat = QIdent "suchthat"
-
--- logical constants in BaseConstants.dk
-propFalse = EIdent identFalse
-propAnd x y = EApp (EApp (EIdent identConj) x) y
-propOr x y = EApp (EApp (EIdent identDisj) x) y
-propImp x y = EApp (EApp (EIdent identImpl) x) y
-propEquiv x y = EApp (EApp (EIdent identEquiv) x) y
-propNot x = EApp (EIdent identNot) x
-
-propPi kind pred = EApp (EApp (EIdent identPi) kind) pred
-propSigma kind pred = EApp (EApp (EIdent identSigma) kind) pred
-
--- built-in types
-typeProp = EIdent identProp
-typeSet = EIdent identSet
-typeType = EIdent identType 
-
-identProp = QIdent "Prop"
-identSet = QIdent "Set"
-identType = QIdent "Type"
-
---- needed for typing conclusions of proofs
-expTyped x t = EApp (EApp (EIdent (QIdent "typed")) x) t
-expNegated x = EApp (EIdent identNeg) x
-
 -- lookup after annotation from dynamically loaded file
 lookupConstantFull :: String -> Maybe (String, String, [String], Int)  -- last arg: drop
 lookupConstantFull f = case splitConstant f of
@@ -200,18 +128,3 @@ escapeConstant s = case s of
 isIdentChar :: Char -> Bool
 isIdentChar c = or [isAlpha c, isDigit c, elem c ".'_"]
 
--- Dedukti representation of digits in BaseConstants.dk
-digitFuns :: [String]
-digitFuns = [nn, nd]
-nn = "nn"
-nd = "nd"
-
-identNn = QIdent nn
-identNd = QIdent nd
-
--- Dedukti representation of lists in BaseConstants.dk
-identNil = QIdent "nil"
-identCons = QIdent "cons"
-
--- Constants related to set in BaseConstants.dk
-identEnumset = QIdent "enumset"
