+++ src/MathCore2Dedukti.hs
@@ -6,6 +6,7 @@ module MathCore2Dedukti where
 import Dedukti.AbsDedukti
 import Informath -- superset of MathCore
 import CommonConcepts
+import DeduktiTheoryAPI
 import DeduktiOperations
 import BuildConstantTable 
 import PGF (showExpr, readExpr, showCId, mkApp, mkCId)
